form-view-model/src/main/kotlin/com/ritense/formviewmodel/service/FormViewModelService.kt
package com.ritense.formviewmodel.service

import com.fasterxml.jackson.databind.ObjectMapper
import com.fasterxml.jackson.databind.node.ObjectNode
import com.ritense.formviewmodel.viewmodel.ViewModel
import kotlin.reflect.KClass

class FormViewModelService(
    val objectMapper: ObjectMapper
) {

    inline fun <reified T : ViewModel> parseViewModel(
        submission: ObjectNode,
        viewModelType: KClass<out T>
    ): ViewModel {
        // When a field is not present in the ViewModel what then? A: it's ignored
        return objectMapper.convertValue(submission, viewModelType.java)
    }

}